[server] Add chunking aware support for A/A producer callback (#131)
This PR fixes the issue that A/A producer callback does not implement ChunkAwareCallback interface. This results in all values will be stored in a single un-chunked value for leader node even if the value produced to VT and saved by follower is chunked. It makes RocksDB data not consistent between leader and follower.
This PR fixes the issue by letting the wrapper callback also implement the ChunkAwareCallback API and in the LeaderProducerCallback it also add support for RMD payload if A/A enabled.
Currently only one PROD store has this issue, we can re-push to create a new version with consistent state when this fix is deployed.