[server] Added a fix for WC chunking data loss
When the output of a WC op would be above the chunking threshold, there
was a bug where the leader would write the whole thing to VT and thus
to followers, but then would persist only the first chunk, and not the
remaining chunks nor the manifest. This data loss in the leader could
then lead to contaminating the VT and followers if another WC op came
in for the same key, as the starting value of the WC would be stale.
The fix for this is to avoid performing DIV on messages that don't have
a directly corresponding upstream message. This is achieved by adding a
new API: LeaderProducedRecordContext::hasCorrespondingUpstreamMessage.
That class is refactored a bit to have stronger ctor param validation
and less mutability. Another code path which already did skip some
bookkeeping in the same scenario is now refactored to use the new API
as well.