[server][common] Sync offset metadata during graceful shutdown (#137)
* [server][common] Sync offset metadata during graceful shutdown

This commits changes the behavior of how OffsetRecord updates its offset metadata. Prior to this, OffsetRecord always sync the metadata from incoming records. Now, it will sync metadata from PartitionConsumptionState.

Due to the design of sync byte interval, the metadata of OffsetRecords will only be synced when consumed bytes have reached the sync byte interval. For stores that have small records, it causes the issue that servers
have to re-ingest large amount of data after a restart because the offsets for these partitions haven't been persisted yet due to not meet the threshold above.

With this change, PartitionConsumptionState will pass through information to OffsetRecords to sync up with while retain the old behavior to avoid aggresive sync-up.
In addition, before graceful shutdown, the metadata of OffsetRecord will also be updated to avoid re-ingesting everything.