[server] Enhance isolated ingestion request handling logic (#151)
During the rolling out of isolated ingestion in EI enviorment, an issue was detected: Each server handles large amount of requests from different stores and drop partition command can be time consuming (it needs to potentially stop consumption and drain records and then close RocksDB) and taking up available threads (currently 10) in the thread pool. New requests can be piling up and eventually time out for the default 60s timeout. The problem can be more severe when the request is actually processed (partition removed from II host and cleaned up the status from II server but does not reset it in main process due to the fact it thinks the request time out and still retrying). The next retry will get rejected as II thinks this topic partition resources does not being to II process. It will stuck in this retry loop and the helix OFFLINE->DROPPED transition will never be completed.

Based on the observation, this PR makes the following tweaks: (1) Introduce a new config to control the timeout and increase the default client request timeout from 60s to 120s (2) Move the main process resource status cleanup logic up to before sending out request to II process so main process should not get stuck in the loop forever.
This PR serves as the short term solution. For long term solution we will need to refactor the request handling logic.