[build] Split integration tests into 18 buckets
The buckets are controlled by a map defined in:

internal/venice-test-common/build.gradle

The build can now also dynamically generate the GitHub CI workflow file
according to the content of the build, by running:

./gradlew generateGHCI

This reduces the wall-clock time of the build to roughly half an hour.

Also overrode Gradle's logging logic to print the runtime of successful
tests.