[server] Cache RT offset lag for lag metric
ready_to_serve_with_rt_lag metrics will try to evaluate RT offset lag between latest persisted offset and RT topic end offset. The call to fetch RT topic end offset is expensive and can lead to slow down of the overall metrics retrieval and results in metrics not displayed correctly on InGraph dashboard.

To solve this issue, this PR proposes to build a region level lag cache map in the PartitionConsumptionState for metrics computation so it won't be slow down.

Before ready to serve is announced, each RT message consumption will trigger lag measurement, so it will also update the cache.
After ready to serve is announced, we will periodically meausre RT offset lag in all regions and update the cache so metrics will still be able to get the updated view.

There is an edge case to be solved here: If the RT does not has new messages to be consumed after a certain period, then the after ready-to-serve call path will not be triggered, and the lag might not become zero even if we consume the last batch of data from RT in the last 1 minute period.