[server] Main process should throw exception when isolated ingestion process throws exception during the ingestion
During the rollout of isolated ingestion we found that IsolatedIngestionBackend swallows the exception thrown in isolated ingestion process during ingestion. This does not align with the non-II behavior and will let server stuck in infinitive retry when there is a zombie resource in Helix that does not exist in ZK.
This PR fixes it by adding a new state in ingestion task report to declare there is exception thrown during the action and main process should also throw it.