[server] Async update CachedKafkaMetadataGetter to avoid metric fetch slowdown (#175)
This PR makes CachedKafkaMetadataGetter value update logic async so metric call to fetch the offset won't be slowdown due to slowness in Kafka metadata fetching.
This PR will issue async call to update value when it found the cached value TTL has expired. All the call to fetch the value will return immediately except when the cache is empty.