[da-vinci][server] fixed an ingestion race condition
When active/active or write-compute is enabled, the consumer thread could
read the local database, while the drainer threads could be adjusting the storage partition
when receiving some special Control Messages/signals, such as StartOfPush or ReadyToServe
signal.
Previously, we coded a fix to make sure the partition id list is always valid for the above
race condition, but that fix couldn't make sure the partition will be accessible in
consumer thread when the same partition is being adjusted, such as close and reopen.

This fix introduces an additonal safeguard to make sure the partition being accessed
will be always valid and it also updates the locking granularity to be partition-level.