[test] Added component tag to the log messages
When running integration tests against a whole cluster or multiple-cluster or two-layer of multiple clusters,
it is very hard to differentiate which log lines are produced by which component, which makes it very challenging
to debug test failures.
This code change is leveraging a feature from log4j2: Inheritable Thread Context to log an unique component id in
each log line, and we could group logs by component tag to figure out most of the logs produced by a certain component.

Component tag examples in the logs:
ParentController-58481
dc-0-ChildController-58413
dc-0-VeniceRouter-58441
dc-0-VeniceServer-58460

The format is {colo_name}-{service_name}-{port}

This is still a best-effort since not all the logs can be tagged with this strategy.
For example, for the logs produced in common thread pool, the component tag will be missing.
Such as the logic executed in `CompletableFuture#runAsync(Runnable runnable)`.