[server][vpj] Calculate target partition before adding chunkedKeySuffix in VeniceWriter::delete
Currently, when chunking is enabled, for DELETE op records we calculate the target partition
after adding a chunkedKeySuffix. This suffix addition to the key changes the partition on
which the record lands. For PUT ops we calculate the target parition before adding any suffix
to the key. This means that the records with the same key but different operation may end up
on different partitions.

To send records to the proper partition we need to get target partition without chunkedKeySuffix
addition to the key.