[server][vpj] Calculate target partition before adding chunkedKeySuff… (#187)
[server][vpj] Calculate target partition before adding chunkedKeySuffix in VeniceWriter::delete (#183)

Currently, when chunking is enabled, for DELETE op records we calculate the target partition after adding a chunkedKeySuffix. However, for PUT ops we calculate the target partition before adding any suffix to the key. 
The suffix added to the key changes the partition on which the record lands. This means that the records with 
the same key but different operations may end up on different partitions. To send records to the proper partition
we need to get the target partition without chunkedKeySuffix addition to the key.

Also made cosmetic changes to the Kafka topic dump tool to make it easier to match logs and topic dump.