[server] Bump ingestion isolation metric request timeout to 60 seconds
We found that ingestion isolation metric request cannot completed within 5s timeout in recent releases debugging.
After frankenWar experiment we found that it takes ~15s to complete in venice-1 EI environment.
This PR adjusts the request timeout as a short term solution to make sure metrics will be collected successfully.
In the long term, we should find a solution to collect the metrics efficiently with much less query time.