[controller][common] Improve listStorePushInfo api to report partition and replica details
Today, listStorePushInfo is an API of parent controller to describe the push job information
of a store's current version in each colo e.g. push job start(end) time. However, it does not
contain per partition/replica raw stats used for estimating the data recovery time.

This rb modifies the current API to have a new flag indicating if partition and replica details
should be returned. If it is set to false, it remains today's behaviour. However, if it is set
to true, it returns replicas details for each composing partition of the store's current version.
This rb also adds a new integration test to verify the new API's correctness.