[controller] Add resilience to pushjob when LEADER replica errors out (#19)
[controller] Add resilience to pushjob when LEADER replica errors out.

If during ingestion LEADER replicas transitions to error state in CV, currently the pushjob will fail without any type of resiliency. To overcome this single point of failure, instead of immediately marking the job failed, we disable the given replica, and let helix choose another replica which hopefully could finish the ingestion without any error. We will try to do this only if we have enough replica available.

The disable partitions are re-enabled by listening to live instance monitor. Whenever the error replica is restarted, the controller will listen to that event and enable the replica in helix. This way server will try to reingest again.

Co-authored-by: smaji <smaji@users.noreply.github.com>