[vpj] Added a secondary sorting for KIF repush to reduce the memory usage in reducer
Without MR secondary sorting, the KIF reducer needs to cache all the messages belonging
to the same key, which includes regular put/delete and chunks and manifest to figure out
the latest value, which is not sustainable for some hot keys, which have been updated many
times, and another known issue to make it worse is that Venice doesn't cleanup the previous
chunks when there is a new update, since the backend will just replace the chunk manifest
entry, but leave alone the previous chunks.
With secondary sort, the assembly of chunked messages is much simpler and cost efficient since
manifest is always with a higher offset than the corresponding chunks, and there is no need
to cache all the decoded messages any more.