[da-vinci][server] Got rid of bounded memory queue for ingestion exceptions
Previously, StoreIngestionTask is using a memory-bounded queue to buffer
the ingestion exception, which is working fine for most of the scenarios,
but it could block StoreBufferService if the exception queue is full without
draining, and it happened when some StoreInestionTask is closed, but there
are still some pending messages in the drainer queue, which would throw a lot
of exceptions to stuck the SToreBufferService since the exception queue is full.

The fix is to use Map to store the last exception per partition, which would never block.
In the meantime, this code change also fixed a couple of flaky tests in StoreIngestionTaskTest.
Also this code change modified the way how to track the ingestion exception per partition:
1. Record the ingestion exception whenever there is an exception thrown.
2. Unsubscribe the errored partition to avoid the data gap issue or useless consumption.
3. Only cleanup the partition exception during rebalance.