[controller] Fixed bug where RT creation in some code path did not respect RF and minISR configs
Currently, there are multiple places a RT (real-time) topic can be created in Venice. One of the
code path did not respect the replication factor and minimum in sync replica config for RT topics.

Out of scope: ideally we should also try to refactor the code so topic creation is centralized and
not all over the place.

Fix default for kafka replication factor