[admin-tool] Implement an admin-tool command to trigger data recovery execution
This change adds a new admin-tool command to start repush-based data recovery workflow.
The new command supports taking multiple stores (or all user stores from a cluster if specified) as input.
Internally, it will run a KIF repush task (an Azkaban push job) for each of the input stores in parallel.
After the mutiple tasks are started, it collects and parse the responds from the Azkaban to show the results.