[server] Force refresh schema when valid superset schema is not found in the schema data (#205)
This PR address the superset schema not found issue during value schema evolution. When user updates value schema, superset schema id might be changed but the corresponding schema might not be refreshed in schema repo at the exact same time. This can lead to LFSIT failed to find the superset schema to deserialize incoming update record and thus fail the ingestion.

This PR fixes it by adding force refresh if it found that superset schema is not in the repo.