[vpj] Change config name from parent.controller.colo.name to parent.controller.region.name (#214)
In PR #195, we started with "colo" to denote data centers. Felix suggested that "colo" was not an industry term and we decided to change it to "region". I changed it in most places, but now realize that I missed one config name. Since this change is pretty new, changing the config to align with the industry.