[server][controller][compat] Add Venice store views concept to controller and server and implement ChangeCaptureViewWriter (#138)
* [controller] WIP Transmit versionSwap message when version changes

We change versions under one of three cases.  Push completion (so long as version swap isn't deferred)
version setting (makeCurrent) or rollback.  We'll only transmit the version swap message if one of the
versions we're transiting to is hybrid, and as a temporary measure, only transmit the message if the
store has a view set up for either version.  This is perhaps temporary, but this removes the blast radius
until we've made changes to make this all relevant for other venice stores.

* [server] Process VersionSwap Message on Servers

EOM

* [server] Add View concept to ingestion task and the logic for changeCaptureViewWriter

* [server][controller][compat] For every record ingested on nearline, signal all views for the store of the event.

This 'should' be all the server side changes needed for Venice change capture views.

* [server] tweak interface to accomodate both currentSchemaId and newSchemaId

EOM

* [server] Make process record block on the venice writer call

EOM

* [tests] Remove extra arguments from storeIngestionTask

Originally had added views as a constructor argument, but decided to break the cycle of interface abuse mid implementation

* Some bug fixes and interface tweaks

* clean up comments

* Adding more tests and some bug fixes

* Address comments and add more test coverage as well as address some test failures

* Adding unit tests to make coverage go up

* Unit test coverage improvements

* Remove keySchema field to make static analysis happy

We can add it back when/if we need it.  It's very likely to be needed for other view implementations but it's not particularly needed for the Change Capture view.  So nixing it for now.

* Added more unit tests

* Added one more unit test and addressed PR comments.

* Remove * imports

---------

Co-authored-by: Zac Policzer <zpolicze@zpolicze-mn1.linkedin.biz>