[server] Add resiliency to isolated ingestion remote requests
This PR adds resiliency to isolated ingestion remote request. In case of regression or extreme request errors, a topic partition may be consdiered ingesting in remoted process but it might actually in main process. This can be blocking server shutdown if it happens to STOP_CONSUMPTION command.

To tackle this, this PR adds a simple logic to check if local ingestion service contains this resource when remote request is rejected. It will make sure in the extreme case, server will be able to shutdown successfully.