[da-vinci][fast-client] Fix stale meta system store metadata in DaVinciClientMetaStoreBasedRepository (#235)
The old metaSystemStoreChangeListener implementation to refresh metadata for meta system store only
works if the repository is also subscribed to the corresponding user store. This is usually the case
for a DVC use case but not in the fast-client use case. In FC we only subscribe to the store's meta
system store and not the actual user store. Meaning the periodic refresh will do nothing after we
bootstrap the corresponding meta system store's current version at the time of initialization.
This is because we stored meta system stores in a separate map and the subscribedStoreMap in
NativeMetadataRepository will be empty in the FC use case.

Fix:
1. On subscribe we also add meta system stores to the subscribedStoreMap in NativeMetadataRepository
so subsequent refresh store will be triggered for meta system stores.

2. Remove metaSystemStoreChangeListener and override refreshOneStore in DaVinciClientMetaStoreBasedRepository
to check and handle any meta system store version updates.