[vpj][test] Support active-active heartbeat store in VPJ. (#171)
Heartbeat store of VPJ (native replication) has been disabled in PROD during CY, as there is a safeguard inside parent controller for disabling request stream topic. To re-enable this heartbeat system store for VPJ, we decided to enable active-active for this store.

Instead of relying on the requesting topic logic to get `VersionCreationResponse` when creating `PushJobHeartbeatSender`, we rely on parent controller client to obtain this store information.