[controller] Fix a race condition between kill job and version swap
Today there's a race around kill job and push completion or version swap. This is because we only write
the kill message and do nothing and rely on the servers to receive the kill and report ERROR to mark
a version killed and prevent it to become current/active. However, this is not always the case especially
when the push strategy only requires n-1 replicas to be completed in every partition. This means if a kill
is sent when a push is near complete or almost complete, it can still become ONLINE and current version
after the kill is sent and recognized by some SNs. This is an issue because we now have a current version
that also have a kill message in the participant store. No re-balance or re-bootstrap can succeed for this
store version.

This fix includes:
1. On kill, we take a store lock (since current version swap will also be acquiring this lock) when
   checking if the version we are trying to kill is still in a non-complete state. Update the version
   status on ZK to be KILLED.
2. On version swap, we check if the version we are swapping to has been killed or not by looking for
   the same flag that we update on the version.
3. An integration test with mocks to verify the correctness of the above behaviour.