[vpj] dict compression for empty push to the hybrid store
For many real-time only stores, it is quite common to schedule an empty push
to deprecate the stale entries.
This code change adds the dict compression support for the empty push to the
hybrid stores.

In the meantime, this code change also fixes a bug in the existing KIF dict compression
support, and it will decompress the payload in the source topic if the compression is
enabled.