[venice-server] Augment ready to serve check to include ignored offsets
Previous to this patch we only took into account the offsets of messages consumed from RT which were committed to local storage.  This did not include messages which were screened due to failing DCR checks. This has the potential race condition of causing rewind to fail should the last events in a given rewind be offsets which we ignored.  This patch makes it so that the we will consider ignored offsets when making this calculation.