[server] Isolated ingestion server should resume LEADER state when restarting forked ingestion process upon crash (#270)
When isolated ingestion process crash, we do have detection and restart mechanism but we do not keep leader status so when resuming ongoing TP ingestion, we will not resume the LEADER partition replica state correctly and LEADER partition could stuck forever.
This PR fixes it by keeping track of the LEADER status and resend STANDBY->LEADER for those LEADER replicas during restart of the forked process. No matter it was in LEADER or in transition to LEADER, this message will let it go through the promotion so it could become LEADER again.