[controller][test] Fix unknown push job details schema error in controller
When a new 'PushJobDetails.avsc' value schema is defined and evolved, there is an issue
that controllers do not load the avsc file and register the new schema, thus causing
errors in VPJ jobs when the VPJ sends a serialized job details using the new schema, but
leader controller fails to get the value schema id because the new schema is
still not registered.

Normally, SystemSchemaInitializationRoutine automatically does the above loading and
registration steps after STANDBY -> LEADER transition is completed. However, for
the PUSH_JOB_DETAILS system store, the corresponding SystemSchemaInitializationRoutine
is missing and this change fix it by adding one for the PUSH_JOB_DETAILS.

An integration test is added to verify the schemas are registered in the controller's
ZK schema repo.