[samza] Augment venice-system-producer to transmit write compute operations when batch type is BATCH
The system producer when encountering a write compute record during a batch write will convert
the write compute operation to a full put. This makes sure that all records in VT are fully specified
and users are able to work with partial schemas when transmitting data during reprocessing from
stream.