[da-vinci][fast-client][test] DaVinci and fast client compatibility fix (#253)
The main issue is that when using DaVinci based store repository there is a cyclic dependency
between DaVinciBackend and the DaVinci baased store repository. This can cause double initialization
of the singleton DaVinciBackend which results in multiple attempts of RocksDB open operation.

The fix is to introduce a new thin-client based metadata and deprecate the DaVinci based metadata. FC
must use the thin-client based metadata if it's used with DVC in the same instance/application.

These are meant to be short term fixes. The long term solution is to remove the DVC dependency from
fast-client and all these problems will go away.