[test] Fix flaky testProducerCloseDoesNotLeaveAnyFuturesIncomplete (#285)
In the test, we want to simulate a case in which there are some records in the producer's buffer that  
are not yet sent to Kafka when an ungraceful close is issued. However, when producer::close() is not   
run before unblocking the producer's ioThread, records are sent to Kafka, which is not what we want. 
To prevent sending records to Kafka, this PR added a check that guarantees producer::close() is  
invoked before unblocking ioThread.