[controller][router][server][dvc] Improved schema mismatch error message
When trying to get the schema ID for a value schema or derived schema
which does not exist, the error message would say that the derived
schema does not exist. This can be confusing if write compute is not
enabled, since derived schemas should not exist in that case.

In this commit, we are making the error messages more precise.

Other clean ups:

- Eliminated a fairly pervasive use of Pair<Integer, Integer> for
  derived schema IDs, and replaced it with a dedicated class.

- Eliminated the unnecessary getZkStoreName static utility function.

- Eliminated some unnecessary lookups in SchemaData.

- Simplified locking and eliminated unnecessary lookups in
  HelixReadOnlySchemaRepository.