[admin-tool] Add a data recovery monitoring command in admin-tool
This change adds a command in admin-tool to monitoring the progresses of one/multiple
triggered data recovery job(s). If the recovery job is still ongoing, it reports the
progress in terms of uncomplelted partitions. This is done by leveraging the newly
improved job status API which returns the status of uncompleted partitions and replicas.
Since we have the raw data of the uncompleted ones, we could further improve the reporting
part to make it more readable to users.

This change also refactors the existing class hierachy to make it easy to add new commands
in the future. DataRecoveryWorker and Command are introduced as new base classes to
abstract out the common logic, so that subclasses can be simplified and focus more on
their special logic.