[controller] Fixed Store config JSON encoding
We used to have a rmdVersionId field in the Store config, with a type of
Optional<Integer>, but this did not get encoded properly by Jackson.

Starting in this commit, in order to avoid breaking on previously-encoded
data, we are renaming the field to rmdVersion, and changing its type to
just int, with -1 indicating that it is unset.

Also renamed the config: replication.metadata.version.id to
replication.metadata.version so that it is consistent with the code.