[server] Align nullable collection field behavior in A/A + Partial Update with existing non-AA logic (#287)
Current RMD/PartialUpdate schema generation consider [null, map] and [null, list] as legit collection field, same as simple map/list, so we need to handle it properly, especially for the NULL value.
Existing non-AA partial update logic treat null as a special type of empty collection:
When setting field to null, it does set the actual value to null. If there is another collection merge operation (add/remove entry), it will first initialize an empty collection and perform collection merge.

This PR address the feature gap by introducing the same logic to A/A PartialUpdate:
(1) When setting a field to null, if it is in put only state (no active elements), it will set to null.
(2) When setting a field to null, if it is not in put only state (has active elements), it will initialize new empty collection and apply active elements.
(3) When a nullable collection field is null, new collection merge operation will also initialize new empty collection and apply new operations.