[dvc][server] Metrics clean up
Eliminated a bunch of metrics of dubious value.

Also replaced the total-only metrics in HostLevelIngestionStats with LongAdders,
wrapped inside a new class called LongAdderRateGauge. If this works well, we can
consider adopting it elsewhere as well, but we need to keep in mind that the
memory footprint of LongAdder grows proportionally to contention and can be
quite high in the worst case, so it may never be appropriate to use this for
more numerous metrics, like per-store ones.

Also refactored some of the DIV code to eliminate a hot spot... according to
profiling, the remaining portion of the DIV code is ~20% of the CPU of what
the whole thing used to take.