[server] Augment ready to serve check to include ignored offsets (#271)
* [venice-server] Augment ready to serve check to include ignored offsets

Previous to this patch we only took into account the offsets of messages consumed from RT which were committed to local storage.  This did not include messages which were screened due to failing DCR checks. This has the potential race condition of causing rewind to fail should the last events in a given rewind be offsets which we ignored.  This patch makes it so that the we will consider ignored offsets when making this calculation.

* Add test cases to improve coverage for change (was required after removing an optional)

* Add in integration test for the scenario.

* remove unused variable so static analysis doesn't get mad

* Tweak when we increment the updateLatestRTOffsetToVTMap to just after we start produce not when we receive an ack back from Kafka

* augment test and add VT lag check for followers

* Add case for empty topic and add some more unit tests

* fix unit test to account for extra method call

---------

Co-authored-by: Zac Policzer <zpolicze@zpolicze-mn1.linkedin.biz>