[fc][tc][dvc][router][server][controller][samza] System store free metadata (#276)
Adding a new metadata system that uses the Venice Server's metadata endpoint instead of the meta system store. We perform cluster discovery using the router's existing endpoint. This response will also contain the server's d2 service name which we can use to hit the metadata endpoint. From there, the metadata is cached in-memory.

Then, to keep the metadata up-to-date, there is a periodic refresh of the metadata by hitting the endpoint and updating the local cache if there are any changes. In the case of store migrations, we must perform re-discovery.

This PR also introduces a Helix Assisted Routing implementation for the Fast Client by taking advantage of the helix group info that the metadata endpoint provides