[fc] retry for cluster discovery (#317)
Added retry for cluster discovery via router in RequestBasedMetadata. The current implementation has this set to false which is resulting in flakiness in the jenkins test runs.

Also added a non-deterministic wait which uses a thin client to query a key. This will ensure D2 and the router are running before the tests run.