[server] A/A + WC should handle update on key that were written in the batch phase (#318)
This PR fixes a bug that MergeConflictResolver.update() will throw exception when it sees oldValue without a RMD.
The assumption is wrong because in the batch phase, AASIT call LFSIT logic to process batch data without creating RMD value. The fix is to find the correct writer schema for the original value and create the new field-level RMD schema.