[dvc][server] Reduced hot path calls to the clock in metrics
Removed ~30 calls to the clock / ingested message.

Fixed a concurrency correctness issue in DIVStats' manually-maintained counters
and changed success and duplicate message counters into LongAdders, since they
are expected to be hot path counters. Also added the code to reset the success
msg counter, which was missing (all other counters in this class had it...).

Changed consumed_record_end_to_end_processing_latency so that it gets recorded
on every message. Previously, it was recorded only when DCR determined that the
write should be dropped, which seemed incorrect.

Removed following metrics:

- leader_delegate_real_time_record_latency (consistently very low, didn't seem
  useful)
- current_idle_time (unused)
- overall_idle_time (unused)