[test][server] Add the ability to plug in different pub-sub systems for integration tests (#309)
This is the first part of a change that will allow us to plug in different pub-sub systems for integration testing. 

To support the new pub-sub system the following two interfaces need to be implemented:
* PubSubBrokerFactory - A factory to create pub-sub service
* PubSubBrokerWrapper - Concrete implementation should encapsulate the logic required to start   
  and stop a pub-sub service.  

We'll pass the `PubSubClientsFactory` object obtained from `PubSubBrokerWrapper` to Venice components such as venice-server to let them know the type of pub-sub clients to use when running tests. 

Added system property `pubSubBackendFactory` to specify which pub-sub system to use for the integration test run.    
If this property is not specified, tests will use the default `KafkaBrokerFactory`.