[controller][vpj][changelog][admin-tool] Shutdown D2 client if it fails startup (#324)
We saw a case when rolling D2ClientFactory out where the D2 client startup failed to start up. In this instance, it didn't clean up the partial resources it had acquired. Some threads inside the D2 client held up the shutdown of the application. To force release of resources, we explicitly shut down the client if it fails during start up