[samza] Adapt modified records in UpdateBuilderImpl to the specified update schema using default values (#323)
The original intention of creating the UpdateBuilder abstraction was that the users need not be aware of other fields in the schema when making partial updates to records. Hence, it is possible that a framework generates the base UpdateBuilder object and the user code only fills in the fields they wish to update. In such a case, the schema of the value object might be different from what is used by the UpdateBuilderImpl object. This is possible if sub fields are evolved, but the writer is still only writing using older schemas.

To reduce the number of user disruptions, the UpdateBuilderImpl will make a best effort to adapt the records specified by the user to the schema that the UpdateBuilderImpl object uses. This is achieved by:
1. Filling in default values for fields that are present in the update schema but not present in the user-provided object
    i. The new fields must have default values since the records must be fully compatible.
2. Removing extra values that are not in update schema
    i. The fields that are not present in the update schema, but are present in the user-specified object, will not be written to Venice.
    ii. From the user's perspective, this is a case of "data loss", but it can be avoided by the framework/application
    iii. To reduce the occurrence of such best-effort adaptations, the application/framework must try to always use the latest superset schema to create the UpdateBuilderImpl object