[thin-client] Add an API to fetch the current superset schema for a store (#327)
After PR #323, when writing partial update values that do not conform to the schema specified in the UpdateBuilderImpl object, the values will be adapted to the schema specified; by filling default values for missing fields and dropping additional fields.
In the case where fields are dropped, from the user's perspective, this is a data-loss. To minimize such cases, the application/framework must always use a superset schema to create the UpdateBuilderImpl object. This commit adds an API on the routers to fetch the superset schema and a corresponding function in the StoreSchemaFetcher interface.