[build] Try to reduce CI job failures by running Gradle Wrapper validation only once (#328)
We have seen CI jobs fail frequently due to SSL errors while running the Gradle wrapper validation tasks. This could be due to the increased load on the server where the validation is run against. Since the use of the validation is only to validate if the gradle-wrapper.jar is not a trojan horse, we can run the check only once