[changelog-client] Add chunking and compression support to change capture client
We add to the change capture client compression and chunking by utilizing a local
storage engine thats held in memory.  A client will only consume events in serial per
partition, and since we only have one producer at a time, we'll only buffer one message
at a time.  Zombie writes may cause the client to leak buffer space temporarily, so
we make a point to continously drop and clear out the storage engines as we receive
new events to keep memory footprint low.

Also added some config validation to creating views. Chunking is now a mandatory
config for the change capture view (which shouldn't be too arduous since we already
require it for enabling write compute which is the most likely case for change capture
to be enabled).