[thin-client] Async retry for store initialization when it fails
This code change added an infinite retry of store initialization when
the first round of blocking initalization fails.
The infinite retry will happen in the background, and the request will fail
before the async thread can successfully initialize the client.
This is a resilience improvement and in theory, this behavior shouldn't happen
often.