[thin-client] Fixed a race condition of thin-client store init retry
In a previous resilience enhancement, there is a race condition that
the store init retry thread could spin up many threads if the store init
keeps failing with a lot of concurrent requests.
This code change guarantees there is only one thread initialized for
the retry logic to avoid zoomie threads.