[dvc][server] Fixed LARG bug in HostLevelIngestionStats
The LARG metrics used for total-only stats in HostLevelIngestionStats were
flat-lining at zero after a regression in PR #300. This refactoring makes
those stats work correctly again, and adds several unit tests at multiple
layers in order to verify that.

Moreover, the LARG class is moved to the venice-client-common module, so
that AbstractVeniceStats can check if a stat belongs to this class, and if
so, rename it to Rate, in order to avoid breaking dashboards based on
existing metric names.

Also removed the Kafka interface from Venice's TestMockTime, and instead
encapsulated that interop in the integration tests module.